# [12.x] perf: Optimize BladeCompiler