@browner12 2025-05-12
is the bulk of the benefit from removing the `in_array` call that was running N times?

I'm curious if you would have got the same improvements sticking with the chained Collection calls, but removing the `in_array`.