@calebdw 2025-05-13
You might could get some additional performance gains by using a variable instead of repeatedly accessing the array:

```php
        foreach (token_get_all($contents) as $token) {
            $token = $token[0];
            if ($token === T_OPEN_TAG || $token === T_OPEN_TAG_WITH_ECHO || $token === T_CLOSE_TAG) {
                $tokens[] = $token;
            }
        }
```