@browner12 2025-05-13
been programming PHP for 20 years and TIL `array_column` exists 🤯