@rzv-me 2025-05-13
@browner12 the bulk of the imprvements come from 2 places:
- not doing the pluck before filtering, even better when you filter and `pluck(0)` at the same time
- not doing the `in_array()` when filtering, as the options are static and very limited (3) 

@calebdw 
- the first suggestion might be a bit better, but it also means there is an extra memory allocation
- as for the second suggestion, I suspect it might be the same as, doing the `pluck(0)` before the filter
did you benchmark them?