@shaedrich 2025-05-13
> ```php
>         foreach (token_get_all($contents) as $token) {
>             $token = $token[0];
>             if ($token === T_OPEN_TAG || $token === T_OPEN_TAG_WITH_ECHO || $token === T_CLOSE_TAG) {
>                 $tokens[] = $token;
>             }
>         }
> ```

You can even avoid using indices altogether by using array destructuring:

```diff lang=php
        foreach (token_get_all($contents) as $token) {
-           $token = $token[0];
+           [$token] = $token;
            if ($token === T_OPEN_TAG || $token === T_OPEN_TAG_WITH_ECHO || $token === T_CLOSE_TAG) {
                $tokens[] = $token;
            }
        }
```